switch to using a fixed hash key all across the runtime

+++ src/err.c
@@ -145,7 +145,7 @@ void output_err(ArErr *err) {
     if (file) {
       dye_style(stderr, DYE_STYLE_RESET);
       dyefg(stderr, DYE_RESET);
-      int line_number_width = snprintf(NULL, 0, "%" PRIu64, frame->line);
+      uint64_t line_number_width = snprintf(NULL, 0, "%" PRIu64, frame->line);
       char *buffer = NULL;
       size_t size = 0;
       uint64_t current_line = 1;
@@ -158,7 +158,7 @@ void output_err(ArErr *err) {
         current_line++;
       }
       fprintf(stderr, "  ");
-      for (int i = 0; i < line_number_width; i++) {
+      for (uint64_t i = 0; i < line_number_width; i++) {
         fprintf(stderr, " ");
       }
       fprintf(stderr, "|\n");
